Implement AC2 and AC3
