+++ src/components/App/App.js
@@ -7,22 +7,25 @@ import Footer from "../Footer/Footer";
 import Header from "../Header/Header";
 import Main from "../Main/Main";
 
-import { useFetchMoviesData } from "../../hooks/useFetchMoviesData";
-
 import "./App.css";
 
 const App = () => {
   const isLoading = useSelector((state) => state.ui.isLoading);
 
-  useFetchMoviesData();
-
   return (
     <ErrorBoundary>
       <Routes>
-        <Route path="/" element={<Navigate to="/search" />}></Route>
-        <Route path="/search" element={<Header />} />
+        <Route path="/" element={<Navigate to="/search?search=" />}></Route>
+        <Route
+          path="search"
+          element={
+            <>
+              <Header />
+              <Main isLoading={isLoading} />
+            </>
+          }
+        />
       </Routes>
-      <Main isLoading={isLoading} />
       <Footer />
     </ErrorBoundary>
   );
